+++ Changes.md
@@ -12,6 +12,7 @@ Version 3.1.1.0
   9. Page -> Background text colour so that setting is more descriptive.
  10. Navbar -> Theme text colour so that the setting is more descriptive.
  11. General tidy up of styles.
+ 12. Add background image settings.
 
 Version 3.1.0.3
   1. Change in architecture a liitle for child themes.